requirements.txt
streamlit
pandas
plotly
requests
fpdf2
markdown
xhtml2pdf
pycoingecko
openai
matplotlib
beautifulsoup4
python-dotenv
google-genai
